refactor emulator
