+++ custom_components/meross_lan/emulator/mixins/garagedoor.py
@@ -4,9 +4,8 @@ from __future__ import annotations
 import asyncio
 import typing
 
-from custom_components.meross_lan.merossclient import const as mc, get_element_by_key
-
-from ..emulator import MerossEmulator
+from .. import MerossEmulator
+from ...merossclient import const as mc, get_element_by_key
 
 
 class GarageDoorMixin(MerossEmulator if typing.TYPE_CHECKING else object):
