+++ custom_components/meross_lan/emulator/mixins/thermostat.py
@@ -3,9 +3,8 @@ from __future__ import annotations
 
 import typing
 
-from custom_components.meross_lan.merossclient import const as mc, get_element_by_key
-
-from ..emulator import MerossEmulator
+from .. import MerossEmulator
+from ...merossclient import const as mc, get_element_by_key
 
 
 class ThermostatMixin(MerossEmulator if typing.TYPE_CHECKING else object):
